Add additional index creation error checking

+++ api/config.py
@@ -148,19 +148,20 @@ for schema_filepath in glob.glob(schema_path + '/input/*.json'):
 assert input_schemas == expected_input_schemas, '{} is different from {}'.format(input_schemas, expected_input_schemas)
 
 def create_or_recreate_ttl_index(coll_name, index_name, ttl):
-    index_list = db[coll_name].index_information()
-    if index_list:
-        for index in index_list:
-            # search for index by given name 
-            # example: "timestamp_1": {"key": [["timestamp", 1]], ...}
-            if index_list[index]['key'][0][0] == index_name:
-                if index_list[index].get('expireAfterSeconds', None) != ttl:
-                    # drop existing, recreate below
-                    db[coll_name].drop_index(index)
-                    break
-                else:
-                    # index exists with proper ttl, bail
-                    return
+    if coll_name in db.collection_names():
+        index_list = db[coll_name].index_information()
+        if index_list:
+            for index in index_list:
+                # search for index by given name 
+                # example: "timestamp_1": {"key": [["timestamp", 1]], ...}
+                if index_list[index]['key'][0][0] == index_name:
+                    if index_list[index].get('expireAfterSeconds', None) != ttl:
+                        # drop existing, recreate below
+                        db[coll_name].drop_index(index)
+                        break
+                    else:
+                        # index exists with proper ttl, bail
+                        return
     db[coll_name].create_index(index_name, expireAfterSeconds=ttl)
 
 
@@ -174,7 +175,7 @@ def initialize_db():
     db.acquisitions.create_index('session')
     db.acquisitions.create_index('uid')
     db.acquisitions.create_index('collections')
-    create_or_recreate_ttl_index('authtokens', 'timestamp', 604800)
+    create_or_recreate_ttl_index('authtokens', 'timestamp', 60)
     create_or_recreate_ttl_index('uploads', 'timestamp', 60)
     create_or_recreate_ttl_index('downloads', 'timestamp', 60)
